Move bitcoin price to lib.rs
