+++ mutiny-wasm/src/lib.rs
@@ -1288,7 +1288,7 @@ impl MutinyWallet {
     /// Gets the current bitcoin price in chosen Fiat.
     #[wasm_bindgen]
     pub async fn get_bitcoin_price(&self, fiat: Option<String>) -> Result<f32, MutinyJsError> {
-        Ok(self.inner.node_manager.get_bitcoin_price(fiat).await?)
+        Ok(self.inner.get_bitcoin_price(fiat).await?)
     }
 
     /// Exports the current state of the node manager to a json object.
